vulnops/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "aegis-vulnops"
version = "0.1.0"
description = "Autonomous vulnerability lifecycle management agent — DefectDojo and Excel sources"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["security", "vulnerability", "defectdojo", "devSecOps", "automation"]

dependencies = [
    "aiohttp>=3.9.5",
    "asyncssh>=2.14.2",
    "gitpython>=3.1.43",
    "jinja2>=3.1.4",
    "openpyxl>=3.1.2",
    "pydantic>=2.7.1",
    "anthropic>=0.28.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
aegis-vulnops = "vulnops.agent:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
]

[tool.setuptools.packages.find]
where = [".."]
include = ["vulnops*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
